requirements.txt
# Core ML
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0

# Experiment Tracking & Versioning
mlflow>=2.9.0
dvc>=3.30.0
dagshub>=0.3.0

# API
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# Monitoring
evidently>=0.4.0

# Utilities
python-dotenv>=1.0.0
joblib>=1.3.0
